finish re-organising files
