+++ __tests__/components/GuideCardOverview.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render } from "@testing-library/react";
-import { GuideCardOverview } from "../../app/components/guideCard/GuideCardOverview";
+import { GuideCardOverview } from "../../app/components/guideCardOverview/GuideCardOverview";
 import {
   ReturnStatus,
   FeedbackStatus,
