+++ __tests__/components/PeoplePage.test.tsx
@@ -1,5 +1,5 @@
 import { render } from "@testing-library/react";
-import PeoplePage from "../app/pages/people/page";
+import PeoplePage from "pages/people/page";
 import { getUsers } from "serverActions/getUsers";
 
 jest.mock("serverActions/getUsers", () => ({
