+++ app/components/guideCard/style.tsx
@@ -1,8 +1,7 @@
 "use client";
 import styled from "styled-components";
-import { Wrapper } from "globalStyles/input/style";
+import { Wrapper } from "UIcomponents/input/style";
 import { Button } from "globalStyles/buttons/default/style";
-import { StyleColors } from "globalStyles/colors";
 
 const BREAKPOINT = "680px";
 const BREAKPOINT_DESKTOP = "1024px";
@@ -27,120 +26,6 @@ export const InfoWrapper = styled.div<{ $borderStyle: string | undefined }>`
   position: relative;
 
   ${(props) => props.$borderStyle}
-`;
-
-export const Info = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  height: 100%;
-  align-items: center;
-  gap: 8px;
-  padding: 20px 0;
-`;
-
-export const GuideNr = styled.h2`
-  font-size: 16px;
-  font-weight: 600;
-  align-self: center;
-`;
-
-export const Name = styled.p`
-  font-size: 12px;
-  width: 155px;
-  font-weight: 400;
-  text-align: center;
-`;
-
-export const ReturnStatusContainer = styled.div`
-  display: flex;
-  gap: 4px;
-  align-items: center;
-`;
-
-export const ModalWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-  max-width: 1200px;
-  width: 100%;
-`;
-
-export const Header = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-  width: 100%;
-  padding-bottom: 32px;
-  @media (min-width: ${BREAKPOINT}) {
-    flex-direction: row;
-  }
-`;
-
-export const TitleContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-  width: 100%;
-`;
-
-export const ColouredCircle = styled.div<{ $backgroundColor?: StyleColors }>`
-  border-radius: 50%;
-  ${(props) => `background-color: ${props.$backgroundColor};`}
-
-  width: 12px;
-  height: 12px;
-
-  @media (min-width: ${BREAKPOINT}) {
-    width: 16px;
-    height: 16px;
-  }
-`;
-
-export const OverviewWrapper = styled(Wrapper)`
-  gap: 2rem;
-  align-self: center;
-  width: 100%;
-`;
-
-export const InfoContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-`;
-
-export const Link = styled.a`
-  text-decoration: none;
-`;
-
-export const ReturnButton = styled(Button)`
-  text-align: left;
-  padding: 2px 8px;
-  white-space: nowrap;
-`;
-
-export const ReturnLinksWrapper = styled.div`
-  display: flex;
-  gap: 8px;
-  flex-wrap: wrap;
-`;
-
-export const StyledLink = styled(Link)`
-  text-decoration: none;
-  color: inherit;
-  display: flex;
-  justify-content: center;
-  align-items: flex-start;
-  width: 100%;
-  height: 100%;
-
-  &:hover {
-    background-color: var(--primary-black-10);
-  }
-`;
-
-export const LinkNoWrap = styled(Link)`
-  white-space: nowrap;
 `;
 
 export const WriteFeedbackContainer = styled.form`
@@ -179,15 +64,4 @@ export const VoteIcon = styled.div`
   height: 100px;
   justify-content: center;
   align-items: center;
-`;
-
-export const GuideDescription = styled.div`
-  display: flex;
-  flex-direction: column;
-  flex: 1;
-  justify-content: center;
-  text-align: center;
-  align-items: flex-start;
-  text-align: center;
-  gap: 0.25rem;
 `;
