+++ app/components/guideModal.tsx/calculateReturnStyle.ts
