+++ app/components/profile/profile.tsx
@@ -13,8 +13,8 @@ import {
   ProfileInfo,
   Logout,
 } from "./style";
-import Modal from "../../UIcomponents/modal/modal";
-import Input from "../../globalStyles/input";
+import Modal from "UIcomponents/modal/modal";
+import { Input } from "UIcomponents/input/Input";
 import DefaultButton from "../../globalStyles/buttons/default";
 import { DefaultUserIcon } from "assets/Icons";
 import { UserType } from "models/user";
