+++ app/components/returnOverview/style.tsx
@@ -1,0 +1,15 @@
+import { Wrapper } from "globalStyles/globalStyles";
+import styled from "styled-components";
+import { Button } from "globalStyles/buttons/default/style";
+
+export const OverviewWrapper = styled(Wrapper)`
+  gap: 2rem;
+  align-self: center;
+  width: 100%;
+`;
+
+export const ReturnLinksWrapper = styled.div`
+  display: flex;
+  gap: 0.5rem;
+  flex-wrap: wrap;
+`;
