+++ app/pages/login/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import LoginOrRegister from "../../UIcomponents/LoginOrRegister";
+import { LoginOrRegister } from "components/loginOrRegister/LoginOrRegister";
 
 export default function LoginPage() {
   return <LoginOrRegister />;
